+++ src/MafiaProject/Member.php
@@ -40,32 +40,6 @@ class Member
         return $this->subordinates;
     }
 
-    public function goToJail()
-    {
-        $this->isFree = false;
-        $newBoss = $this->getSusbstituteSameRange();
-        if (is_null($newBoss)) {
-            $newBoss = $this->promoteSubordinate();
-        }
-        $this->addSubordinatesToNewBoss($newBoss);
-    }
-
-    protected function getSusbstituteSameRange()
-    {
-        $posiblesBosses = $this->getBoss()->getSubordinates();
-        $maxAge = 0;
-        /** @var Member $newBoss */
-        $newBoss = null;
-        /** @var Member $posible */
-        foreach ($posiblesBosses as $posible) {
-            if ($posible->getAge() > $maxAge && $posible->getIsFree()) {
-                $maxAge = $posible->getAge();
-                $newBoss = $posible;
-            }
-        }
-        return $newBoss;
-    }
-
     public function getBoss()
     {
         return $this->boss;
@@ -86,21 +60,9 @@ class Member
         return $this->isFree;
     }
 
-    protected function promoteSubordinate()
+    public function setIsFree($isFree)
     {
-        $posiblesBosses = $this->getSubordinates();
-        $maxAge = 0;
-        $newBoss = null;
-        /** @var Member $posible */
-        foreach ($posiblesBosses as $posible) {
-            if ($posible->getAge() > $maxAge) {
-                $maxAge = $posible->getAge();
-                $newBoss = $posible;
-            }
-        }
-
-        $this->getBoss()->addSubordinate($newBoss);
-        return $newBoss;
+        $this->isFree = $isFree;
     }
 
     public function addSubordinate(Member $member)
@@ -109,23 +71,9 @@ class Member
         $member->setBoss($this);
     }
 
-    protected function addSubordinatesToNewBoss(Member $newBoss)
-    {
-        foreach ($this->getSubordinates() as $subordinateToMove) {
-            /** @var $subordinateToMove Member */
-            if (strcmp($subordinateToMove->getName(), $newBoss->getName()) !== 0) {
-                $newBoss->addSubordinate($subordinateToMove);
-            }
-        }
-    }
-
     public function getName()
     {
         return $this->name;
     }
 
-    public function isImportantMember()
-    {
-        return $this->getNumberOfSubordinates() > 50;
-    }
 }
